requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==1.8.2
torch==1.9.0
torchvision==0.10.0
pillow==8.3.1
python-dotenv==0.19.0
efficientnet-pytorch>=0.7.1
opencv-python>=4.8.0
numpy>=1.24.3
scikit-learn>=1.3.0
pymongo>=4.5.0
jinja2>=3.1.2
aiofiles>=23.1.0
requests>=2.31.0
matplotlib>=3.7.2
pandas>=2.0.3

